fix: added shebang in the setup.sh
This commit is trivial. It is to append the shebang in the setup.sh file. 

It is not always the case that /bin/bash is the path to the Bash shell.
It is only a different path in extremely rare instances, so it is typically irrelevant.
This commit is to specify the Bash shell (i.e., bash) using the environment variable.

Signed-off-by: Geunsik Lim <leemgs@gmail.com>