feat: add option of launch.sh
This PR is to add an option function for the user to selectively use
foreground mode and background mode (i.e., daemon mode).
Also, depending on the version of docker composer, the -d (daemon mode) option
may or may not be present. Therefore, this is to deal effectively with these
situations.

Signed-off-by: Geunsik Lim <leemgs@gmail.com>
Signed-off-by: Geunsik Lim <geunsik.lim@samsung.com>